Former-commit-id: 1b4447e046a1b42be46aeac2bc4d06a5d19a07cc

+++ CommonLib/Threadpool.h
@@ -14,7 +14,7 @@ public:
 	void CleanUp();
 	
 	bool HasWork() { return !myWorkOrders.empty(); }
-	int GetWorkSize() const { return myWorkOrders.size(); }
+	size_t GetWorkSize() const { return myWorkOrders.size(); }
 	bool CurrentWorkFinished() const;
 private:
 	std::queue<Work> myWorkOrders;
