+++ Engine/TerrainManager.h
@@ -8,11 +8,11 @@ class TerrainManager
 public:
 	TerrainManager() = default;
 	bool CleanUp();
-	Terrain* GetTerrain(u32 terrain_hash);
+	Terrain* GetTerrain(u64 terrain_hash);
 	Terrain* GetTerrain(const std::string& terrain_name);
 private:
 	bool m_FinishedLoading = false;
-	std::map<u32, Terrain*> m_Terrain;
-	void LoadTerrain(u32 terrain_hash);
+	std::map<u64, Terrain*> m_Terrain;
+	void LoadTerrain(u64 terrain_hash);
 	void LoadTerrain(const std::string& terrain_name);
 };
